Downloader
A server-client based project (group of 4) developed using python in
which a central server is connected to multiple
clients. On the request of a client, the server downloads a particular
file using multiple threads to enhance the
downloading speed and then transfers the file to the client. This file
is now also stored at the server so that the
request for the same file by any other client can be served instantly.